refactor: adapater article detail
